WIP continued, Joy split into files
